+++ src/config.h
@@ -316,6 +316,6 @@ extern char * save_filename(char * suf);
 // In a configuration parameter list, map tag to value
 extern char * matchconf(char * conf, char * item);
 extern wchar * config_log;
-extern char * keyclick;
+extern void * keyclick;
 
 #endif